MAIN_SDK_CH32V307_V1/Components/simulation/software/requirements_pyqt.txt
PyQt5>=5.15
pyqtgraph>=0.13
pyserial>=3.5

